# Switch to multi-context seeds hash function