--- a/src/cmdline.hpp
+++ b/src/cmdline.hpp
@@ -52,6 +52,7 @@ struct CommandLineOptions {
     int u { 7 };
     int s { 16 };
     int c { 8 };
+    int aux_len{24};
 
     // Alignment
     int A { 2 };