@ksahlin 2024-10-01
If I understand correctly, the only change in this PR is to use

```
    if (hash1 > hash2) {
        std::swap(hash1, hash2);
    }
    return ((hash1 >> aux_len) << aux_len) ^ (hash2 >> (64 - aux_len));
```
instead of `return hash1 + hash2;` as the randstrobe hash?

And you observe a slight speedup from this change? I guess it could happen if  `return hash1 + hash2` creates some collisions? (should be very rare). It doesn't seem that `+` would be more expensive than the other version by itself..