@marcelm 2024-10-02
> If I understand correctly, the only change in this PR is to use
> 
> ```
>     if (hash1 > hash2) {
>         std::swap(hash1, hash2);
>     }
>     return ((hash1 >> aux_len) << aux_len) ^ (hash2 >> (64 - aux_len));
> ```
> 
> instead of `return hash1 + hash2;` as the randstrobe hash?

Yes, that’s the only relevant change.

> And you observe a slight speedup from this change? I guess it could happen if `return hash1 + hash2` creates some collisions? (should be very rare). It doesn't seem that `+` would be a more expensive operation than the other version by itself..

Yes, the addition is a single machine instruction whereas the new function has more than 10 instructions or so.

I just realized that this is probably faster because the randstrobes in the index are ordered in such a way that makes us use the cache better. Let’s say you have a query with syncmers A, B, C and strobemers are paired up A-C, B-C. If C has the lowest hash value of all three, it will be the main hash in both cases. Since entries in the index are primarily sorted by main hash, entries for A-C and B-C are very close in the index. So after A-C has been looked up, B-C is likely already in the cache.

I used `perf` to measure this

Old hash function:
```
Performance counter stats for '...':

    14,694,060,241      cache-references                                                   
     1,670,029,387      cache-misses                     #   11.365 % of all cache refs    
   752,100,535,203      cycles                                                             
 1,863,007,791,894      instructions                     #    2.48  insn per cycle         
   260,485,816,961      branches                                                           

      43.124037916 seconds time elapsed

     167.170358000 seconds user
       0.343364000 seconds sys
```

New hash function:
```
Performance counter stats for '...':

    14,960,029,504      cache-references                                                   
     1,403,230,152      cache-misses                     #    9.380 % of all cache refs    
   733,275,948,339      cycles                                                             
 1,872,996,951,027      instructions                     #    2.55  insn per cycle         
   261,331,524,007      branches                                                           

      42.163663935 seconds time elapsed

     163.632210000 seconds user
       0.379490000 seconds sys
```

So the number of instructions, cycles and branches all go up with the new function, which is what we expect, but the number of cache misses goes down, which is what in this case has the greatest effect.