@ksahlin 2024-10-02
I see, clever! `perf` is something I should make use of.