@marcelm 2024-10-02
> I see, clever! `perf` is something I should make use of.

It’s a Linux-only tool, but there’s probably something similar for macOS.

> Would it then make sense to sort the query seeds by hash before lookups, or will sorting and 'desorting' for find_nams (if needed) kill the gains?

For an individual query, probably not so much. Since there are relatively few randstrobes per query, you will still read from very different parts of memory even if you sort by hash. Accessing monotonically increasing memory addresses per se doesn’t help; an advantage exists only when you access the same cache line more than once. (Cache lines are something like 64 bytes large.)

Maybe obtaining all randstrobes for a large number of reads beforehand and then sorting them by hash could give an advantage.

We could also use prefetching (as suggested in #203).