@marcelm 2024-10-02
> The number of times the same main is seen by chance (as currently) depends on the window parameters (thus read length). But I guess the fraction of same main occurring twice in a row is very low, particularly for longer reads.

Note that it doesn’t have to be consecutive: Even if the second access is somewhere downstream, it’s still likely in the cache.

> For the shortest it may be more common as window starts at the immediate downstream window. Did you higher speedup for the shorter reads?

I cannot say because the speedup is just 1-4%, so it’s hard to measure.

> Extending the thought: Then we might as well sort FW and RC seeds together before lookup - increasing expected repetitive stretch of same main strobe to four.

That could work, but it’s probably not necessary as I’m guessing that the entries in the index from the forward pass over the query are probably still in the cache during the reverse pass.

Also, sorting will probably eat up any benefits (I have a data point for this in my next PR).