@ksahlin 2024-10-03
Adding here a note from email conversation: 

Sorting on hash value would also mean we don’t have to store a partial look up vector/set which is a commit that increase the runtime with about 6% (https://github.com/ksahlin/strobealign/pull/426#issuecomment-2379082142). In that case a variable holding the previous main hash would be sufficient.