@marcelm 2024-10-03
I don’t think this will help cache efficiency, but not having to keep track of partial lookups would reduce runtime by a couple of percent.

I think we could even do this without any extra cost: We already need to sort the *matches* by query coordinate in `merge_hits_into_nams`. If we instead sort the *hits* in `find_nams` (as you suggest), we shouldn’t have to sort again.