@marcelm 2024-10-03
Hm, you meant sorting by hash value, but I meant something slightly different. I’ll think about it.